requirements.txt
boto3==1.26.16
botocore==1.29.16
click==8.1.3
colorama==0.4.6
dominate==2.7.0
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
Pillow==9.3.0
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.13
visitor==0.1.3
Werkzeug==2.2.2
zipp==3.10.0